+++ src/commands/apphosting-secrets-describe.ts
@@ -9,7 +9,7 @@ import { requirePermissions } from "../requirePermissions";
 import * as Table from "cli-table3";
 
 export const command = new Command("apphosting:secrets:describe <secretName>")
-  .description("Get metadata for secret and its versions.")
+  .description("get metadata for secret and its versions")
   .before(requireAuth)
   .before(secretManager.ensureApi)
   .before(requirePermissions, ["secretmanager.secrets.get"])