+++ src/commands/dataconnect-sql-diff.ts
@@ -9,7 +9,7 @@ import { requireAuth } from "../requireAuth";
 
 export const command = new Command("dataconnect:sql:diff [serviceId]")
   .description(
-    "displays the differences between  a local DataConnect schema and your CloudSQL database's current schema",
+    "display the differences between a local Data Connect schema and your CloudSQL database's current schema",
   )
   .before(requirePermissions, [
     "firebasedataconnect.services.list",