+++ src/commands/firestore-backups-list.ts
@@ -9,10 +9,10 @@ import { logWarning } from "../utils";
 import { PrettyPrint } from "../firestore/pretty-print";
 
 export const command = new Command("firestore:backups:list")
-  .description("List all Cloud Firestore backups in a given location")
+  .description("list all Cloud Firestore backups in a given location")
   .option(
     "-l, --location <locationId>",
-    "Location to search for backups, for example 'nam5'. Run 'firebase firestore:locations' to get a list of eligible locations. Defaults to all locations.",
+    "location to search for backups, for example 'nam5'. Run 'firebase firestore:locations' to get a list of eligible locations. Defaults to all locations",
   )
   .before(requirePermissions, ["datastore.backups.list"])
   .before(warnEmulatorNotSupported, Emulators.FIRESTORE)