+++ src/commands/firestore-databases-delete.ts
@@ -13,10 +13,8 @@ import { FirebaseError } from "../error";
 import { PrettyPrint } from "../firestore/pretty-print";
 
 export const command = new Command("firestore:databases:delete <database>")
-  .description(
-    "Delete a database in your Cloud Firestore project. Database delete protection state must be disabled. To do so, use the update command: firebase firestore:databases:update <database> --delete-protection DISABLED",
-  )
-  .option("--force", "Attempt to delete database without prompting for confirmation.")
+  .description("delete a database in your Cloud Firestore project")
+  .option("--force", "attempt to delete database without prompting for confirmation")
   .before(requirePermissions, ["datastore.databases.delete"])
   .before(warnEmulatorNotSupported, Emulators.FIRESTORE)
   .action(async (database: string, options: FirestoreOptions) => {